chore: reliable way to start amman + test validator
- works as long as tests are run in series as otherwise all of them
  start amman in parallel

+++ client-rust/src/test_utils/mod.rs
@@ -1,13 +1,17 @@
-use lazy_static::lazy_static;
 use std::{
-    io,
-    process::{Child, Command},
+    io::{self, Read},
+    net::TcpStream,
+    process::{Child, Command, Stdio},
+    str,
 };
 use thiserror::Error;
 
 use crate::blocking::AmmanClient;
 
 pub type AmmanProcessResult<T> = Result<T, AmmanProcessError>;
+
+// TODO(thlorenz): may have to get this from env in the future
+const AMMAN_EXECUTABLE: &str = "amman_";
 
 #[derive(Error, Debug)]
 pub enum AmmanProcessError {
@@ -24,19 +28,23 @@ pub enum AmmanProcessError {
     FailedToKillAmman(#[from] io::Error),
 }
 
-lazy_static! {
-    pub static ref AMMAN: AmmanProcess = {
-        let client = AmmanClient::new(None);
-        let mut amman = AmmanProcess::new(client);
-        amman.ensure_started().unwrap();
-        amman
-    };
-}
-
 pub struct AmmanProcess {
     process: Option<Child>,
     pid: Option<u32>,
     client: AmmanClient,
+}
+
+impl Clone for AmmanProcess {
+    fn clone(&self) -> Self {
+        // Cannot clone the process, thus this mainly serves to not have to query the pid
+        // for an externally running amman again.
+        // It is mainly used when attempting to restart the validator.
+        Self {
+            process: None,
+            pid: self.pid.clone(),
+            client: self.client.clone(),
+        }
+    }
 }
 
 impl AmmanProcess {
@@ -66,13 +74,36 @@ impl AmmanProcess {
         if let Some(pid) = pid_of_amman_running_on_machine(&self.client) {
             return Err(AmmanProcessError::AmmanAlreadyRunning(pid));
         }
-        let process = Command::new("amman_").arg("start").spawn()?;
-        while pid_of_amman_running_on_machine(&self.client).is_none() {}
+
+        let mut cmd = Command::new(AMMAN_EXECUTABLE);
+        cmd.arg("start").stdout(Stdio::null()).stderr(Stdio::null());
+        let process = cmd.spawn()?;
+
+        eprint!("\nWaiting for pid");
+        loop {
+            match pid_of_amman_running_on_machine(&self.client) {
+                Some(pid) => {
+                    eprintln!(": {:#?}", pid);
+                    break;
+                }
+                None => {}
+            }
+        }
+
+        eprint!("Waiting for validator to be ready: ");
+        wait_for_port(8900);
+        eprint!("✔️\n");
         self.process = Some(process);
         Ok(())
     }
 
-    pub fn kill(&mut self) -> AmmanProcessResult<()> {
+    pub fn restart(&mut self) -> AmmanProcessResult<()> {
+        self.kill(true)?;
+        self.start()?;
+        Ok(())
+    }
+
+    pub fn kill(&mut self, kill_external: bool) -> AmmanProcessResult<()> {
         if self.process.is_none() && self.pid.is_none() {
             return Err(AmmanProcessError::AmmanCannotBeKilledIfNotRunning);
         }
@@ -86,7 +117,13 @@ impl AmmanProcess {
             process.wait()?;
             self.process = None;
         } else if let Some(pid) = self.pid {
-            eprintln!("Refusing to kill process that was not created by this runner ({:#?}). Please kill via `amman stop`",  pid);
+            if kill_external {
+                let mut process = Command::new(AMMAN_EXECUTABLE).arg("stop").spawn()?;
+                process.wait()?;
+                self.pid = None;
+            } else {
+                eprintln!("Refusing to kill process that was not created by this runner ({:#?}). Please kill via `amman stop`",  pid);
+            }
         }
 
         Ok(())
@@ -113,6 +150,17 @@ pub fn pid_of_amman_running_on_machine(client: &AmmanClient) -> Option<u32> {
         Ok(pid) => Some(pid),
         Err(_) => None,
     }
+}
+
+fn scan_port(port: u16) -> bool {
+    match TcpStream::connect(("0.0.0.0", port)) {
+        Ok(_) => true,
+        Err(_) => false,
+    }
+}
+
+fn wait_for_port(port: u16) {
+    while !scan_port(port) {}
 }
 
 /*
